refactor(website): extract PipelineVisual card components into components/
Move OpportunityCard and its parts (FieldRow, CompanyChip, DealPersonChip,
CreatedByChip) into PipelineVisual/components/, one file each, with the
card-level styled co-located in their consuming component. Shared deal
types move to deal-types.ts. PipelineVisual.tsx now keeps only the board
shell, data, and drag orchestration.
